+++ src/basic_java_class.cpp
@@ -26,6 +26,9 @@ namespace RexVM {
             JAVA_LANG_RUNTIME_NAME,
             JAVA_LANG_INVOKE_METHOD_HANDLE_NAME,
             JAVA_LANG_STACK_TRACE_ELEMENT_NAME,
+            JAVA_LANG_REFLECT_METHOD_NAME,
+            JAVA_LANG_REFLECT_FIELD_NAME,
+            JAVA_LANG_INVOKE_MEMBER_NAME_NAME,
     };
 
     cstring getBasicJavaClassName(BasicJavaClassEnum classEnum) {
