+++ src/config.hpp
@@ -69,6 +69,7 @@ namespace RexVM {
         REF,
     };
 
+
     const Slot ZERO_SLOT = Slot(CAST_I8(0));
 
     inline bool isWideSlotType(SlotTypeEnum slotType) {
@@ -104,32 +105,12 @@ namespace RexVM {
     using MethodHandler = void (*)(Frame &frame);
     using NativeMethodHandler = MethodHandler;
 
-    enum class AccessFlagEnum : u2 {
-        ACC_PUBLIC = 0x0001,
-        ACC_PRIVATE = 0x0002,
-        ACC_PROTECTED = 0x0004,
-        ACC_STATIC = 0x0008,
-        ACC_FINAL = 0x0010,
-        ACC_SUPER = 0x0020,
-        ACC_SYNCHRONIZED = 0x0020,
-        ACC_VOLATILE = 0x0040,
-        ACC_BRIDGE = 0x0040,
-        ACC_TRANSIENT = 0x0080,
-        ACC_VARARGS = 0x0080,
-        ACC_NATIVE = 0x0100,
-        ACC_INTERFACE = 0x0200,
-        ACC_ABSTRACT = 0x0400,
-        ACC_STRICT = 0x0800,
-        ACC_SYNTHETIC = 0x1000,
-        ACC_ANNOTATION = 0x2000,
-        ACC_ENUM = 0x4000,
-        ACC_MODULE = 0x8000,
-        ACC_MANDATED = 0x8000,
-    };
-
     constexpr size_t STRING_POOL_SIZE = 1024;
 
 }
+
+#include "enums.hpp"
+#include "utils/format.hpp"
 
 
 #endif
