+++ src/native/core/java_lang_class.hpp
@@ -6,12 +6,14 @@
 #include "../../thread.hpp"
 #include "../../oop.hpp"
 #include "../../class.hpp"
+#include "../../class_member.hpp"
 #include "../../execute.hpp"
 #include "../../memory.hpp"
 #include "../../constant_info.hpp"
 #include "../../class_loader.hpp"
 #include "../../string_pool.hpp"
 #include "../../exception_helper.hpp"
+#include "../../attribute_info.hpp"
 #include "../../utils/string_utils.hpp"
 #include "../../utils/class_utils.hpp"
