+++ src/print_helper.hpp
@@ -1,18 +1,18 @@
 #ifndef PRINT_HELPER_HPP
 #define PRINT_HELPER_HPP
 
-#include "utils/format.hpp"
-#include "constant_info.hpp"
-#include "attribute_info.hpp"
-#include "class_file.hpp"
-
+#include "config.hpp"
+#include <vector>
+#include <tuple>
 
 namespace RexVM {
 
+    struct ClassFile;
     struct Class;
     struct ObjArrayOop;
     struct Oop;
     struct InstanceOop;
+    struct Frame;
 
     void printConstant(ClassFile &cf);
